Remove distributed caching, only cache for character scoped data, simplify

+++ zone/data_bucket.cpp
@@ -8,7 +8,7 @@
 
 extern WorldServer worldserver;
 
-std::vector<DataBucketCacheEntry> g_data_bucket_cache = {};
+std::vector<DataBucketsRepository::DataBuckets> g_data_bucket_cache = {};
 
 void DataBucket::SetData(const std::string &bucket_key, const std::string &bucket_value, std::string expires_time)
 {
@@ -59,13 +59,12 @@ void DataBucket::SetData(const DataBucketKey &k)
 
 	if (bucket_id) {
 		// loop cache and update cache value and timestamp
-		for (auto &ce: g_data_bucket_cache) {
-			if (CheckBucketMatch(ce.e, k)) {
-				ce.e             = b;
-				ce.updated_time  = GetCurrentTimeUNIX();
-				ce.update_action = DataBucketCacheUpdateAction::Upsert;
-				SendDataBucketCacheUpdate(ce);
-				break;
+		if (CanCache(k)) {
+			for (auto &ce: g_data_bucket_cache) {
+				if (CheckBucketMatch(ce, k)) {
+					ce = b;
+					break;
+				}
 			}
 		}
 
@@ -74,18 +73,8 @@ void DataBucket::SetData(const DataBucketKey &k)
 	else {
 		b.key_ = k.key;
 		b = DataBucketsRepository::InsertOne(database, b);
-		if (!ExistsInCache(b)) {
-			// add data bucket and timestamp to cache
-			auto ce = DataBucketCacheEntry{
-				.e = b,
-				.updated_time = DataBucket::GetCurrentTimeUNIX(),
-				.update_action = DataBucketCacheUpdateAction::Upsert
-			};
-
-			g_data_bucket_cache.emplace_back(ce);
-
-			SendDataBucketCacheUpdate(ce);
-
+		if (CanCache(k) && !ExistsInCache(b)) {
+			g_data_bucket_cache.emplace_back(b);
 			DeleteFromMissesCache(b);
 		}
 	}
@@ -113,21 +102,21 @@ DataBucketsRepository::DataBuckets DataBucket::GetData(const DataBucketKey &k, b
 	);
 
 	for (const auto &ce: g_data_bucket_cache) {
-		if (CheckBucketMatch(ce.e, k)) {
-			if (ce.e.expires > 0 && ce.e.expires < std::time(nullptr)) {
-				LogDataBuckets("Attempted to read expired key [{}] removing from cache", ce.e.key_);
+		if (CheckBucketMatch(ce, k)) {
+			if (ce.expires > 0 && ce.expires < std::time(nullptr)) {
+				LogDataBuckets("Attempted to read expired key [{}] removing from cache", ce.key_);
 				DeleteData(k);
 				return DataBucketsRepository::NewEntity();
 			}
 
 			// this is a bucket miss, return empty entity
 			// we still cache bucket misses, so we don't have to hit the database
-			if (ce.e.id == 0) {
+			if (ce.id == 0) {
 				return DataBucketsRepository::NewEntity();
 			}
 
-			LogDataBuckets("Returning key [{}] value [{}] from cache", ce.e.key_, ce.e.value);
-			return ce.e;
+			LogDataBuckets("Returning key [{}] value [{}] from cache", ce.key_, ce.value);
+			return ce;
 		}
 	}
 
@@ -150,17 +139,14 @@ DataBucketsRepository::DataBuckets DataBucket::GetData(const DataBucketKey &k, b
 			// cache bucket misses, so we don't have to hit the database
 			// when scripts try to read a bucket that doesn't exist
 			g_data_bucket_cache.emplace_back(
-				DataBucketCacheEntry{
-					.e = DataBucketsRepository::DataBuckets{
-						.id = 0,
-						.key_ = k.key,
-						.value = "",
-						.expires = 0,
-						.character_id = k.character_id,
-						.npc_id = k.npc_id,
-						.bot_id = k.bot_id
-					},
-					.updated_time = DataBucket::GetCurrentTimeUNIX()
+				DataBucketsRepository::DataBuckets{
+					.id = 0,
+					.key_ = k.key,
+					.value = "",
+					.expires = 0,
+					.character_id = k.character_id,
+					.npc_id = k.npc_id,
+					.bot_id = k.bot_id
 				}
 			);
 
@@ -184,22 +170,17 @@ DataBucketsRepository::DataBuckets DataBucket::GetData(const DataBucketKey &k, b
 		return {};
 	}
 
-	bool      has_cache = false;
+	bool has_cache = false;
+
 	for (auto &ce: g_data_bucket_cache) {
-		if (ce.e.id == r[0].id) {
+		if (ce.id == r[0].id) {
 			has_cache = true;
 			break;
 		}
 	}
 
 	if (!has_cache) {
-		// add data bucket and timestamp to cache
-		g_data_bucket_cache.emplace_back(
-			DataBucketCacheEntry{
-				.e = r[0],
-				.updated_time = DataBucket::GetCurrentTimeUNIX()
-			}
-		);
+		g_data_bucket_cache.emplace_back(r[0]);
 	}
 
 	return r[0];
@@ -230,7 +211,7 @@ bool DataBucket::DeleteData(const std::string &bucket_key)
 // GetDataBuckets bulk loads all data buckets for a mob
 bool DataBucket::GetDataBuckets(Mob *mob)
 {
-	DataBucketLoadType::Type t;
+	DataBucketLoadType::Type t{};
 	const uint32             id = mob->GetMobTypeIdentifier();
 
 	if (!id) {
@@ -261,14 +242,8 @@ bool DataBucket::DeleteData(const DataBucketKey &k)
 		std::remove_if(
 			g_data_bucket_cache.begin(),
 			g_data_bucket_cache.end(),
-			[&](DataBucketCacheEntry &ce) {
-				bool match = CheckBucketMatch(ce.e, k);
-				if (match) {
-					ce.update_action = DataBucketCacheUpdateAction::Delete;
-					SendDataBucketCacheUpdate(ce);
-				}
-
-				return match;
+			[&](DataBucketsRepository::DataBuckets &ce) {
+				return CheckBucketMatch(ce, k);
 			}
 		),
 		g_data_bucket_cache.end()
@@ -381,13 +356,13 @@ void DataBucket::BulkLoadEntities(DataBucketLoadType::Type t, std::vector<uint32
 		bool            has_cache = false;
 		for (const auto &ce: g_data_bucket_cache) {
 			if (t == DataBucketLoadType::Bot) {
-				has_cache = ce.e.bot_id == ids[0];
+				has_cache = ce.bot_id == ids[0];
 			}
 			else if (t == DataBucketLoadType::Client) {
-				has_cache = ce.e.character_id == ids[0];
+				has_cache = ce.character_id == ids[0];
 			}
 			else if (t == DataBucketLoadType::NPC) {
-				has_cache = ce.e.npc_id == ids[0];
+				has_cache = ce.npc_id == ids[0];
 			}
 		}
 
@@ -442,12 +417,7 @@ void DataBucket::BulkLoadEntities(DataBucketLoadType::Type t, std::vector<uint32
 		if (!ExistsInCache(e)) {
 			LogDataBucketsDetail("bucket id [{}] bucket key [{}] bucket value [{}]", e.id, e.key_, e.value);
 
-			g_data_bucket_cache.emplace_back(
-				DataBucketCacheEntry{
-					.e = e,
-					.updated_time = GetCurrentTimeUNIX()
-				}
-			);
+			g_data_bucket_cache.emplace_back(e);
 		}
 	}
 
@@ -469,11 +439,11 @@ void DataBucket::DeleteCachedBuckets(DataBucketLoadType::Type t, uint32 id)
 		std::remove_if(
 			g_data_bucket_cache.begin(),
 			g_data_bucket_cache.end(),
-			[&](DataBucketCacheEntry &ce) {
+			[&](DataBucketsRepository::DataBuckets &ce) {
 				return (
-					(t == DataBucketLoadType::Bot && ce.e.bot_id == id) ||
-					(t == DataBucketLoadType::Client && ce.e.character_id == id) ||
-					(t == DataBucketLoadType::NPC && ce.e.npc_id == id)
+					(t == DataBucketLoadType::Bot && ce.bot_id == id) ||
+					(t == DataBucketLoadType::Client && ce.character_id == id) ||
+					(t == DataBucketLoadType::NPC && ce.npc_id == id)
 				);
 			}
 		),
@@ -499,144 +469,12 @@ int64_t DataBucket::GetCurrentTimeUNIX()
 bool DataBucket::ExistsInCache(const DataBucketsRepository::DataBuckets &e)
 {
 	for (const auto &ce: g_data_bucket_cache) {
-		if (ce.e.id == e.id) {
+		if (ce.id == e.id) {
 			return true;
 		}
 	}
 
 	return false;
-}
-
-bool DataBucket::SendDataBucketCacheUpdate(const DataBucketCacheEntry &e)
-{
-	if (!e.e.id) {
-		return false;
-	}
-
-	if (e.e.character_id > 0) {
-		return false;
-	}
-
-	EQ::Net::DynamicPacket p;
-	p.PutSerialize(0, e);
-
-	auto pack_size = sizeof(ServerDataBucketCacheUpdate_Struct) + p.Length();
-	auto pack      = new ServerPacket(ServerOP_DataBucketCacheUpdate, static_cast<uint32_t>(pack_size));
-	auto buf       = reinterpret_cast<ServerDataBucketCacheUpdate_Struct *>(pack->pBuffer);
-
-	buf->cereal_size = static_cast<uint32_t>(p.Length());
-
-	memcpy(buf->cereal_data, p.Data(), p.Length());
-
-	worldserver.SendPacket(pack);
-
-	return true;
-}
-
-void DataBucket::HandleWorldMessage(ServerPacket *p)
-{
-	DataBucketCacheEntry         n;
-	auto                         s = (ServerDataBucketCacheUpdate_Struct *) p->pBuffer;
-	EQ::Util::MemoryStreamReader ss(s->cereal_data, s->cereal_size);
-	cereal::BinaryInputArchive   archive(ss);
-	archive(n);
-
-	LogDataBucketsDetail(
-		"Received cache packet for id [{}] key [{}] value [{}] action [{}]",
-		n.e.id,
-		n.e.key_,
-		n.e.value,
-		static_cast<int>(n.update_action)
-	);
-
-	// delete
-	if (n.update_action == DataBucketCacheUpdateAction::Delete) {
-		DeleteFromMissesCache(n.e);
-
-		g_data_bucket_cache.erase(
-			std::remove_if(
-				g_data_bucket_cache.begin(),
-				g_data_bucket_cache.end(),
-				[&](DataBucketCacheEntry &ce) {
-					bool match = n.e.id > 0 && ce.e.id == n.e.id;
-					if (match) {
-						LogDataBuckets(
-							"[delete] cache key [{}] id [{}] cache_size before [{}] after [{}]",
-							ce.e.key_,
-							ce.e.id,
-							g_data_bucket_cache.size(),
-							g_data_bucket_cache.size() - 1
-						);
-					}
-					return match;
-				}
-			),
-			g_data_bucket_cache.end()
-		);
-
-		return;
-	}
-
-	// update
-	bool has_key = false;
-	for (auto &ce: g_data_bucket_cache) {
-		// update cache
-		if (ce.e.id == n.e.id) {
-			// reject old updates
-			int64 time_delta = ce.updated_time - n.updated_time;
-			if (ce.updated_time >= n.updated_time) {
-				LogDataBuckets(
-					"Attempted to update older cache key [{}] rejecting old time [{}] new time [{}] delta [{}] cache_size [{}]",
-					ce.e.key_,
-					ce.updated_time,
-					n.updated_time,
-					time_delta,
-					g_data_bucket_cache.size()
-				);
-				return;
-			}
-
-			DeleteFromMissesCache(n.e);
-
-			LogDataBuckets(
-				"[update] cache id [{}] key [{}] value [{}] old time [{}] new time [{}] delta [{}] cache_size [{}]",
-				ce.e.id,
-				ce.e.key_,
-				n.e.value,
-				ce.updated_time,
-				n.updated_time,
-				time_delta,
-				g_data_bucket_cache.size()
-			);
-			ce.e            = n.e;
-			ce.updated_time = n.updated_time;
-			has_key = true;
-			break;
-		}
-	}
-
-	// create
-	if (!has_key) {
-		DeleteFromMissesCache(n.e);
-
-		size_t size_before = g_data_bucket_cache.size();
-
-		g_data_bucket_cache.emplace_back(
-			DataBucketCacheEntry{
-				.e = n.e,
-				.updated_time = GetCurrentTimeUNIX()
-			}
-		);
-
-		LogDataBuckets(
-			"[create] Adding new cache id [{}] key [{}] value [{}] cache size before [{}] after [{}]",
-			n.e.id,
-			n.e.key_,
-			n.e.value,
-			size_before,
-			g_data_bucket_cache.size()
-		);
-	}
 }
 
 void DataBucket::DeleteFromMissesCache(DataBucketsRepository::DataBuckets e)
@@ -649,11 +487,11 @@ void DataBucket::DeleteFromMissesCache(DataBucketsRepository::DataBuckets e)
 		std::remove_if(
 			g_data_bucket_cache.begin(),
 			g_data_bucket_cache.end(),
-			[&](DataBucketCacheEntry &ce) {
-				return ce.e.id == 0 && ce.e.key_ == e.key_ &&
-					   ce.e.character_id == e.character_id &&
-					   ce.e.npc_id == e.npc_id &&
-					   ce.e.bot_id == e.bot_id;
+			[&](DataBucketsRepository::DataBuckets &ce) {
+				return ce.id == 0 && ce.key_ == e.key_ &&
+					   ce.character_id == e.character_id &&
+					   ce.npc_id == e.npc_id &&
+					   ce.bot_id == e.bot_id;
 			}
 		),
 		g_data_bucket_cache.end()
@@ -672,24 +510,47 @@ void DataBucket::ClearCache()
 	LogInfo("Cleared data buckets cache");
 }
 
-void DataBucket::DeleteCharacterFromCache(uint64 character_id) {
+void DataBucket::DeleteFromCache(uint64 id, DataBucketLoadType::Type type)
+{
 	size_t size_before = g_data_bucket_cache.size();
 
 	g_data_bucket_cache.erase(
 		std::remove_if(
 			g_data_bucket_cache.begin(),
 			g_data_bucket_cache.end(),
-			[&](DataBucketCacheEntry &ce) {
-				return ce.e.character_id == character_id;
+			[&](DataBucketsRepository::DataBuckets &ce) {
+				switch (type) {
+					case DataBucketLoadType::Bot:
+						return ce.bot_id == id;
+					case DataBucketLoadType::Client:
+						return ce.character_id == id;
+					default:
+						return false;
+				}
 			}
 		),
 		g_data_bucket_cache.end()
 	);
 
 	LogDataBuckets(
-		"Deleted character id [{}] from cache size before [{}] after [{}]",
-		character_id,
+		"Deleted [{}] id [{}] from cache size before [{}] after [{}]",
+		DataBucketLoadType::Name[type],
+		id,
 		size_before,
 		g_data_bucket_cache.size()
 	);
+}
+
+
+// CanCache returns whether a bucket can be cached or not
+// characters are only in one zone at a time so we can cache locally to the zone
+// bots (not implemented) are only in one zone at a time so we can cache locally to the zone
+// npcs (ids) can be in multiple zones so we can't cache locally to the zone
+bool DataBucket::CanCache(const DataBucketKey &key)
+{
+	if (key.character_id > 0) {
+		return true;
+	}
+
+	return false;
 }
